# Add 64-byte block loop for purego